Fixed buffer overflow when reading from I2C.

+++ src/bms/eBike_bms.c
@@ -146,7 +146,7 @@ eBike_err_t eBike_bms_read_current(double* pack_current) {
     if (eBike_err.eBike_err_type != EBIKE_OK)
         goto eBike_clean;
 
-    double result = ((cc_register.cc_reading * 8.44) / 1000000) / bq76930_adc_characteristics.shunt_value;
+    double result = ((double) (cc_register.cc_reading * 8.44) / 1000000.0) / bq76930_adc_characteristics.shunt_value;
     memcpy(pack_current, &result, sizeof(double));
     
 eBike_clean:
